Update sessions.py-fixed none problem with saved_session expiry

+++ src/flask_session/sessions.py
@@ -554,7 +554,7 @@ class SqlAlchemySessionInterface(SessionInterface):
         saved_session = self.sql_session_model.query.filter_by(
             session_id=store_id
         ).first()
-        if saved_session and saved_session.expiry <= datetime.utcnow():
+        if saved_session and saved_session.expiry and saved_session.expiry <= datetime.utcnow():
             # Delete expired session
             self.db.session.delete(saved_session)
             self.db.session.commit()
